# Fix `IS NOT NULL` performance regression on object columns