… server/src/main/java/org/elasticsearch/common/lucene/ShardCoreKeyMap.java
… server/src/main/java/org/elasticsearch/index/cache/query/IndexQueryCache.java
… server/src/main/java/org/elasticsearch/index/cache/query/QueryCache.java