fc8ec2f62edfc281515a353fff66a8b481c4b009 led to a performance regression
(https://github.com/crate/crate/pull/13106) because to fix detection of
empty objects it had to add a generic function filter)

This cherry-picks 634ea84c5f20f0d3e4311f5ece1680d7e2ce1f65 which turns
out to have led to a major performance improvement for `IS NOT NULL`
queries:

    V1: 5.2.2-a33862e52a115ac11be517f8f5daf719951a0ad5
    V2: 5.2.4-06e9c3bc2f3eeb6df5ea858335299a5bda310f61
    Q: SELECT count(*) FROM obj_null WHERE o IS NOT NULL;
    C: 1
    | Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
    |   V1    |     1921.848 ± 1828.960 |     94.083 |    839.692 |   4010.403 |   4126.200 |
    |   V2    |        7.708 ±   15.243 |      2.037 |      2.802 |      3.323 |     51.067 |
    ├---------┴-------------------------┴------------┴------------┴------------┴------------┘
    |               - 198.40%                           - 198.67%
    There is a 99.61% probability that the observed difference is not random, and the best estimate of that difference is 198.40%
    The test has statistical significance

    System/JVM Metrics (durations in ms, byte-values in MB)
        |    YOUNG GC            |       OLD GC           |      HEAP         |     ALLOC
        |  cnt      avg      max |  cnt      avg      max |  initial     used |     rate      total
     V1 |  197     1.33     1.27 |    5    21.13    22.99 |     2147      103 | 12566.23     241649
     V2 |    0     0.00     0.00 |    1    47.27    47.27 |     2147      231 |     0.00      13245

It basically negates the performance regression introduced by
fc8ec2f62edfc281515a353fff66a8b481c4b009 and reported in
https://github.com/crate/crate/issues/13769

Comparing with 4.8.4:

    V1: 4.8.4-14d79d0dfa8b978e18e9e1fe07cbbc075f62ed3b
    V2: 5.2.4-44ad95269e0be565be73c5f7e045a88d9cb79255

    Q: SELECT count(*) FROM obj_null WHERE o IS NOT NULL;
    C: 1
    | Version |         Mean ±    Stdev |        Min |     Median |         Q3 |        Max |
    |   V1    |       30.820 ±   85.931 |      2.883 |      3.445 |      4.272 |    275.379 |
    |   V2    |        8.376 ±   16.030 |      2.131 |      3.331 |      4.016 |     53.969 |
    ├---------┴-------------------------┴------------┴------------┴------------┴------------┘
    |               - 114.52%                           -   3.35%
    There is a 57.25% probability that the observed difference is not random, and the best estimate of that difference is 114.52%
    The test has no statistical significance

    System/JVM Metrics (durations in ms, byte-values in MB)
        |    YOUNG GC            |       OLD GC           |      HEAP         |     ALLOC
        |  cnt      avg      max |  cnt      avg      max |  initial     used |     rate      total
     V1 |    0     0.00     0.00 |    0     0.00     0.00 |     2147        0 |  7402.66       1162
     V2 |    0     0.00     0.00 |    0     0.00     0.00 |     2147        0 |     0.00      13404

Original commit message was:

Remove ShardCoreKeyMap and trim IndicesQueryCache

As far as I can tell the ShardCoreKeyMap was effectively unused. The
tracked items were only accessed to clear them again, but not for
anything else.

Given that the Lucene `QueryCache` interface has no close it should be
fine to just let it get GCd