@mfussenegger 2023-03-08
> I'm wondering if ShardCoreKeyMap removal positively affects to genericFunctionQuery usages in general

Very likely, yes